Time: 99 ms (10.47%), Space: 14 MB (77.54%) - LeetHub

+++ 0001-two-sum/0001-two-sum.cpp
@@ -4,13 +4,13 @@ public:
         vector<int> answer;
         answer.reserve(2);
 
-        unsigned short numsSize = nums.size();
+        int numsSize = nums.size();
 
-        for (unsigned short i = 0; i < numsSize - 1; ++i)
+        for (int i = 0; i < numsSize - 1; ++i)
         {
             bool found = false;
 
-            for (unsigned short j = i + 1; j < numsSize; ++j)
+            for (int j = i + 1; j < numsSize; ++j)
             {
                 if (nums[i] + nums[j] == target)
                 {
